==> src/r/1_data_processing/2_process_vector_datasets.qmd <==
---
title: "Process static datasets"
format: html
---

```{r}
#| label: setup
pacman::p_load(tidyverse, sf, terra)

# Load Buffered Bon Bon dataset (from 1_study_area.qmd)
study_area <- read_sf("./outputs/vector/bonbon_buffer.gpkg")
# Get the coordinate extents of the study area
study_extent <- study_area %>% ext()
```

# Stations

```{r}
# Load pastoral stations dataset
stations <- read_sf("./data/vector/pastoral_stations.gpkg") %>%
  st_transform(crs(study_area)) %>% # Project to same coordinate system as study area
  st_crop(., st_bbox(study_area)) %>% # Crop to the bounding box of the study area
  st_as_sf() # Convert to sf

# Write processed pastoral stations data to outputs folder
write_sf(stations, "./outputs/vector/pastoral_stations.gpkg")
```

# IBRA Regions

```{r}
# Read IBRA subregions geopackage
ibra <- read_sf("./data/vector/ibra_subregions.gpkg") %>%
  vect() %>% # Convert to vect (Terra vector format)
  crop(study_extent) %>% # Crop using the study extent
  st_as_sf() # Convert to sf

# Write processed IBRA subregions to outputs folder
st_write(ibra, "./outputs/vector/ibra_subregions.gpkg", delete_layer = T)
```

# Pastoral Landsystems

```{r}
# Load pastoral land systems dataset
pastoral_landsystems <- read_sf("./data/vector/pastoral_landsystems.gpkg") %>%
  vect() %>% # Convert to vect (Terra vector format)
  crop(study_extent) %>%  # Crop using the study extent
  st_as_sf() # Convert to sf

# Write processed landsystens to outputs folder
st_write(pastoral_landsystems, "./outputs/vector/pastoral_landsystems.gpkg", delete_layer = T)
```

# Station Fence Buffer
Create 500 m buffer either side of Bon Bon boundary
```{r}
# Load pastoral stations geodatabase
fences <- read_sf("./outputs/vector/pastoral_stations.gpkg") %>%
  .[which(.$NAME == "Bon Bon"),] %>% # Select Bon Bon only 
  st_transform("EPSG:28353") %>% # Project to local coordinate system
  st_cast("POLYGON") %>% # Convert to polygon
  st_cast("LINESTRING") %>% # Convert to linestring
  st_buffer(500) %>% # Buffer linestring by 500m
  st_union() %>% # Union lines (into single geometry)
  as_tibble() %>% # Convert to tibble
  add_column(fence = 1) %>% # Add column called fence with value of 1
  st_as_sf() %>% # Convert back to sf
  vect() %>% # Convert to vect (Terra vector format)
  crop(study_extent) # Crop geometry to the study extent

# Write Bon Bon fence buffer to outputs folder
writeVector(fences, "./outputs/vector/bonbon_boundary_buffer.gpkg", overwrite = T)
```

# Water
Water polygons, used to mask pixels to exclude

```{r}
# Load geopackage of surface water dataset
water <- read_sf("./data/vector/water_courses.gpkg") %>%
  vect() %>% # Convert to vect
  crop(study_extent) %>% # Crop to study extent
  st_as_sf() %>% # Convert to sf
  st_buffer(500) # Buffer by 250 m to be sure

# Write water areas to outputs folder
write_sf(water, "./outputs/vector/water_courses.gpkg")
```
