==> src/r/1_data_processing/3_process_agcd.qmd <==
---
title: "Process AGCD"
format: html
---

```{r}
#| label: setup
pacman::p_load(terra, tidyverse, stringr, sf)

# Load study area (from 1_study_area.qmd)
studyExtent <- read_sf("./outputs/vector/bonbon_buffer.gpkg") %>% 
  st_geometry()
```


```{r}
# File path to the stored location of agcd netcdf files
filePath <- "E:/data/climate_data/rainfall/agcd/" %>% # Folder location
  list.files("*.nc$", full.names = T) # List .nc files in this folder

rasters <- rast(filePath) # Load rasters from file paths above
# Extract dates from raster times
dates <- time(rasters) %>% as.character()
substr(dates, 9, 10) <- "01" # Replace dates with the 1st

# Rename rasters with monthly dates 
names(rasters) <- basename(filePath) %>% 
  substr(1, 37) %>% 
  paste0(.,"_", dates) %>% 
  str_replace(".nc", "")

# Crop rasters to the study extent (+ 10 km buffer)
rasters <- rasters %>%  crop(studyExtent %>%
                               st_transform("EPSG:4326") %>% # project study extent to match rasters coordinate system 
                               st_buffer(10000)) %>%
  # Reproject to GDA94 MGA zone 53
  project("EPSG:28353") %>% 
  # Crop again to remove edge distortion due to reprojection
  crop(studyExtent %>% 
         st_buffer(5000)) 

dates <- time(rasters)
str_sub(dates, -2) <- "01"
time(rasters) <- as_date(dates)
```

# Export rasters

```{r}
start_date <- time(rasters)[1] %>% 
  str_replace_all("-", "") # Get date of first raster for file name
end_date <- time(rasters)[length(time(rasters))] %>% 
  str_replace_all("-", "") # Get date of last raster for file name

outPath <- filePath %>% 
  basename %>%
  file.path("./outputs/raster/agcd_precipitation", .)

writeCDF(rasters, outPath, varname = "agcd_precipitation_monthly", compression = 6, overwrite = T)
```
